+++ .gitignore
@@ -1,6 +1,6 @@
 .DS_Store
 node_modules
-/src/dist
+/dist
 /demo/dist/*.js
 .env
 yarn-error.log
