+++ demo/src/components/Icons.js
@@ -1,4 +1,4 @@
-import { view } from "../../../src/main";
+import { view } from "../../../src";
 
 export const ThumbsUp = () => view`<svg
     xmlns="http://www.w3.org/2000/svg"
