+++ package.json
@@ -4,6 +4,7 @@
   "main": "dist/viewview.cjs.js",
   "module": "dist/viewview.esm.js",
   "browser": "dist/viewview.umd.js",
+  "types": "dist/index.d.ts",
   "dependencies": {},
   "devDependencies": {
     "@emotion/css": "^11.1.3",
@@ -21,11 +22,11 @@
     "typescript": "^4.2.4"
   },
   "scripts": {
-    "build": "rollup -c",
+    "build": "rm -rf dist/* && rollup -c && tsc --emitDeclarationOnly",
     "dev": "rollup -c rollup.config.dev.js -w",
     "test": "ts-node test/test.ts rollup.config.dev.js",
     "pretest": "yarn build",
-    "publish": "rollup -c rollup.config.js && npm publish --access=public"
+    "publish": "yarn build && npm publish --access=public"
   },
   "files": [
     "dist"
