+++ src/index.ts
@@ -1,0 +1,5 @@
+export * from "./view";
+export { view } from "./view";
+
+export * from "./useViewModel";
+export { useViewModel } from "./useViewModel";
