--- src/main.ts
@@ -1,2 +1,0 @@
-export * as viewview from "./public";
-export { view, useViewModel } from "./public";
