--- src/public/index.ts
@@ -1,2 +1,0 @@
-export * from "./view";
-export * from "./useViewModel";
