--- src/public/modules/index.ts
@@ -1,1 +1,0 @@
-export * from "./schema";
