+++ src/view/external/index.ts
